Reshaping the package structure of the whole game by adding story, dto, scene, configuration packages vol 2
